+++ app/src/main/java/woowacourse/shopping/domain/repository/CartRepository.kt
@@ -5,6 +5,12 @@ import woowacourse.shopping.domain.model.CartItem
 interface CartRepository {
     fun getCartItemCount(onResult: (Result<Int?>) -> Unit)
 
+    fun fetchPagedCartItems(
+        page: Int,
+        pageSize: Int,
+        onResult: (Result<List<CartItem>>) -> Unit,
+    )
+
     fun getTotalQuantity(onResult: (Result<Int?>) -> Unit)
 
     fun insertProduct(
