+++ app/src/main/java/woowacourse/shopping/presentation/cart/CartViewModelFactory.kt
@@ -7,7 +7,6 @@ import woowacourse.shopping.di.RepositoryModule
 class CartViewModelFactory : ViewModelProvider.Factory {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         val cartRepository = RepositoryModule.provideCartRepository()
-        val productRepository = RepositoryModule.provideProductRepository()
-        return CartViewModel(cartRepository, productRepository) as T
+        return CartViewModel(cartRepository) as T
     }
 }
