+++ app/src/main/res/layout/item_cart_product.xml
@@ -32,6 +32,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="6dp"
+            android:text="@{cartItem.product.name}"
             android:textColor="@color/black"
             android:textSize="20sp"
             android:textStyle="bold"
